Refactor parse_format_method_string from strings to utils
