+++ pylint/checkers/utils.py
@@ -33,6 +33,7 @@
 import builtins
 from functools import lru_cache, partial, singledispatch
 import itertools
+import numbers
 import re
 import sys
 import string
@@ -535,65 +536,102 @@ def parse_format_string(
     return keys, num_args, key_types, pos_types
 
 
-def parse_brace_format_string(format_string: str) -> Tuple[Set[str], int]:
-    """Parses a str.format() style format string, otherwise the same as
-    parse_format_string().
+if PY3K:
+    import _string  # pylint: disable=wrong-import-position, wrong-import-order
+
+    def split_format_field_names(
+        format_string
+    ) -> Tuple[str, Iterable[Tuple[bool, str]]]:
+        try:
+            return _string.formatter_field_name_split(format_string)
+        except ValueError:
+            raise IncompleteFormatString()
+
+
+else:
+
+    def _field_iterator_convertor(iterator):
+        for is_attr, key in iterator:
+            if isinstance(key, numbers.Number):
+                yield is_attr, int(key)
+            else:
+                yield is_attr, key
+
+    def split_format_field_names(
+        format_string
+    ) -> Tuple[str, Iterable[Tuple[bool, str]]]:
+        try:
+            keyname, fielditerator = format_string._formatter_field_name_split()
+        except ValueError:
+            raise IncompleteFormatString()
+        # it will return longs, instead of ints, which will complicate
+        # the output
+        return keyname, _field_iterator_convertor(fielditerator)
+
+
+def collect_string_fields(format_string) -> Iterable[Optional[str]]:
+    """ Given a format string, return an iterator
+    of all the valid format fields. It handles nested fields
+    as well.
     """
-    keys = set()  # type: Set[str]
+    formatter = string.Formatter()
+    try:
+        parseiterator = formatter.parse(format_string)
+        for result in parseiterator:
+            if all(item is None for item in result[1:]):
+                # not a replacement format
+                continue
+            name = result[1]
+            nested = result[2]
+            yield name
+            if nested:
+                for field in collect_string_fields(nested):
+                    yield field
+    except ValueError as exc:
+        # Probably the format string is invalid.
+        if exc.args[0].startswith("cannot switch from manual"):
+            # On Jython, parsing a string with both manual
+            # and automatic positions will fail with a ValueError,
+            # while on CPython it will simply return the fields,
+            # the validation being done in the interpreter (?).
+            # We're just returning two mixed fields in order
+            # to trigger the format-combined-specification check.
+            yield ""
+            yield "1"
+            return
+        raise IncompleteFormatString(format_string)
+
+
+def parse_format_method_string(
+    format_string: str
+) -> Tuple[List[Tuple[str, List[Tuple[bool, str]]]], int, int]:
+    """
+    Parses a PEP 3101 format string, returning a tuple of
+    (keys, num_args, manual_pos_arg),
+    where keys is the set of mapping keys in the format string, num_args
+    is the number of arguments required by the format string and
+    manual_pos_arg is the number of arguments passed with the position.
+    """
+    keys = []
     num_args = 0
-
-    def next_char(i):
-        if i + 1 > len(format_string):
-            print("Final: ", format_string[:i])
-            raise IncompleteFormatString
-        return i + 1, format_string[i]
-
-    i = 0
-    while i < len(format_string):
-        i, char = next_char(i)
-        if char == "{":
-            i, char = next_char(i)
-            key = ""
-            # Parse escaped brace character
-            if char == "{":
-                continue
-            # Parse field_name (optional)
-            while char not in "!:}":
-                if char in "[.":
-                    while char not in "!:}":
-                        i, char = next_char(i)
-                    break
-                key += char
-                i, char = next_char(i)
-            # Parse conversion (optional)
-            if char == "!":
-                i, char = next_char(i)
-                if char in "rsa":
-                    i, char = next_char(i)
-                else:
-                    raise UnsupportedFormatCharacter
-            # Parse format_spec (optional)
-            if char == ":":
-                i, char = next_char(i)
-                # ignore the format_spec details
-                while char != "}":
-                    i, char = next_char(i)
-            # Parse end of replacement field
-            if char == "}":
-                if key:
-                    if key not in keys:
-                        num_args += 1
-                        keys.add(key)
-                else:
-                    num_args += 1
-                continue
-            raise UnsupportedFormatCharacter
-        if char == "}":
-            i, char = next_char(i)
-            if char == "}":
-                continue
-            raise UnsupportedFormatCharacter
-    return keys, num_args
+    manual_pos_arg = set()
+    for name in collect_string_fields(format_string):
+        if name and str(name).isdigit():
+            manual_pos_arg.add(str(name))
+        elif name:
+            keyname, fielditerator = split_format_field_names(name)
+            if isinstance(keyname, numbers.Number):
+                # In Python 2 it will return long which will lead
+                # to different output between 2 and 3
+                manual_pos_arg.add(str(keyname))
+                keyname = int(keyname)
+            try:
+                keys.append((keyname, list(fielditerator)))
+            except ValueError:
+                raise IncompleteFormatString()
+        else:
+            num_args += 1
+    return keys, num_args, len(manual_pos_arg)
 
 
 def is_attr_protected(attrname: str) -> bool:
